# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Provision Civils - Construction Job Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs:
    get:
      operationId: listJobs
      tags: [jobs]
      summary: List all jobs
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: supervisorId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
    post:
      operationId: createJob
      tags: [jobs]
      summary: Create a new job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
  /jobs/{id}:
    get:
      operationId: getJob
      tags: [jobs]
      summary: Get a job by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobDetail"
        "404":
          description: Not found

    put:
      operationId: updateJob
      tags: [jobs]
      summary: Update a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Not found

    delete:
      operationId: deleteJob
      tags: [jobs]
      summary: Delete a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /jobs/{id}/photos:
    get:
      operationId: listJobPhotos
      tags: [jobs]
      summary: List photos for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          required: false
          schema:
            type: string
            enum: [before, during, after, other]
      responses:
        "200":
          description: Photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobPhoto"

    post:
      operationId: addJobPhoto
      tags: [jobs]
      summary: Add a photo to a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobPhotoInput"
      responses:
        "201":
          description: Photo added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobPhoto"
  /jobs/{id}/photos/{photoId}:
    delete:
      operationId: deleteJobPhoto
      tags: [jobs]
      summary: Delete a job photo
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: photoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
  /jobs/{id}/photos/zip:
    get:
      operationId: downloadJobPhotosZip
      tags: [jobs]
      summary: Download photos as ZIP (optionally filtered by category)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: category
          in: query
          required: false
          description: Filter by category. Omit to download all photos.
          schema:
            type: string
            enum: [before, during, after, other]
      responses:
        "200":
          description: ZIP archive
          content:
            application/zip:
              schema:
                type: string
                format: binary
        "404":
          description: No photos found
  /jobs/{id}/materials:
    put:
      operationId: updateJobMaterials
      tags: [jobs]
      summary: Update materials checklist for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [materials]
              properties:
                materials:
                  type: array
                  items:
                    $ref: "#/components/schemas/JobMaterialInput"
      responses:
        "200":
          description: Materials updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobMaterial"
  /jobs/{id}/gps-logs:
    get:
      operationId: listJobGpsLogs
      tags: [jobs]
      summary: List GPS logs for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: GPS logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GpsLog"

    post:
      operationId: createJobGpsLog
      tags: [jobs]
      summary: Create a GPS log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GpsLogInput"
      responses:
        "201":
          description: GPS log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GpsLog"
  /jobs/{id}/daily-reports:
    get:
      operationId: listDailyReports
      tags: [jobs]
      summary: List daily reports for a job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Daily reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DailyReport"

    post:
      operationId: createDailyReport
      tags: [jobs]
      summary: Create a daily report
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DailyReportInput"
      responses:
        "201":
          description: Daily report created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
...
components:
  schemas:
...
    Job:
      type: object
      required: [id, jobNumber, clientName, status, createdAt]
      properties:
        id:
          type: integer
        jobNumber:
          type: string
        clientId:
          type: ["integer", "null"]
        clientName:
          type: string
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        projectName:
          type: ["string", "null"]
        projectNumber:
          type: ["string", "null"]
        projectManagerId:
          type: ["integer", "null"]
        poNumber:
          type: ["string", "null"]
        clientOrderNumber:
          type: ["string", "null"]
        contractValue:
          type: ["number", "null"]
        siteAddress:
          type: ["string", "null"]
        gpsLat:
          type: ["number", "null"]
        gpsLng:
          type: ["number", "null"]
        description:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        labourHours:
          type: ["number", "null"]
        status:
          type: string
          enum: [pending, in_progress, active, waiting_for_wayleave, waiting_for_materials, completed, cancelled]
        supervisorId:
          type: ["integer", "null"]
        startDate:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        wayleaveRequired:
          type: boolean
        wayleaveDocument:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    JobInput:
      type: object
      required: [clientName]
      properties:
        clientId:
          type: integer
        clientName:
          type: string
        clientPhone:
          type: string
        clientEmail:
          type: string
        projectName:
          type: string
        projectNumber:
          type: string
        projectManagerId:
          type: integer
        poNumber:
          type: string
        clientOrderNumber:
          type: string
        contractValue:
          type: number
        siteAddress:
          type: string
        gpsLat:
          type: number
        gpsLng:
          type: number
        description:
          type: string
        notes:
          type: string
        labourHours:
          type: number
        supervisorId:
          type: integer
        startDate:
          type: string
        dueDate:
          type: string
        status:
          type: string
          enum: [pending, in_progress, active, waiting_for_wayleave, waiting_for_materials, completed, cancelled]
        wayleaveRequired:
          type: boolean
        workerIds:
          type: array
          items:
            type: integer
        materials:
          type: array
          items:
            $ref: "#/components/schemas/JobMaterialInput"
        equipment:
          type: array
          items:
            $ref: "#/components/schemas/JobEquipmentInput"
...
    JobMaterialInput:
      type: object
      required: [name, quantity, unit]
      properties:
        name:
          type: string
        quantity:
          type: number
        unit:
          type: string
        cost:
          type: number
        checked:
          type: boolean
        notes:
          type: ["string", "null"]
        isCustom:
          type: boolean
...
    JobPhoto:
      type: object
      required: [id, jobId, uri, category, createdAt]
      properties:
        id:
          type: integer
        jobId:
          type: integer
        uri:
          type: string
        caption:
          type: ["string", "null"]
        category:
          type: string
          enum: [before, during, after, other]
        uploadedById:
          type: ["integer", "null"]
        createdAt:
          type: string
    JobPhotoInput:
      type: object
      required: [uri]
      properties:
        uri:
          type: string
        caption:
          type: string
        category:
          type: string
          enum: [before, during, after, other]
...